requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.10.4
asgiref==3.8.1
async-asgi-testclient==1.4.11
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.7
filetype==1.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
loguru==0.7.3
lxml==5.2.2
msgpack==1.1.0
multidict==6.1.0
nonebot-adapter-onebot==2.4.6
nonebot-adapter-qq==1.6.1
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-localstore==0.7.3
nonebot-plugin-send-anything-anywhere==0.7.1
nonebot2==2.4.1
nonebot_plugin_session==0.3.2
nonebug==0.4.3
nonestorage==0.1.0
numpy==2.2.1
packaging==24.2
pillow==10.4.0
pixie-python==4.3.0
pluggy==1.5.0
propcache==0.2.1
psutil==6.1.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygtrie==2.5.0
pypinyin==0.53.0
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qq-botpy==1.2.1
qrcode==8.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.2
StrEnum==0.4.15
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
yarl==1.18.3
